# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Data Room Backend API
  version: 2.0.0
  description: |
    Firebase-authenticated API with Google Drive OAuth integration and
    hierarchical Data Room filesystem endpoints.
servers:
  - url: http://localhost:5000
    description: Local development server
security:
  - bearerAuth: []
tags:
  - name: Health
  - name: User
  - name: Google Integration
  - name: Folders
  - name: Items
  - name: Files
paths:
  /api/health:
    get:
      tags: [Health]
      summary: Health check
      operationId: getHealth
      security: []
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /api/me:
    get:
      tags: [User]
      summary: Get current authenticated user
      operationId: getMe
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /api/integrations/google/status:
    get:
      tags: [Google Integration]
      summary: Get Google Drive integration status
      operationId: getGoogleIntegrationStatus
      responses:
        '200':
          description: Integration status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoogleStatusResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /api/integrations/google/connect:
    post:
      tags: [Google Integration]
      summary: Create OAuth authorization URL for Google Drive connect flow
      operationId: postGoogleConnect
      responses:
        '200':
          description: OAuth authorization URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoogleConnectResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/integrations/google/callback:
    get:
      tags: [Google Integration]
      summary: Google OAuth callback endpoint
      operationId: getGoogleCallback
      security: []
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: error
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to frontend OAuth callback route
          headers:
            Location:
              schema:
                type: string
                format: uri

  /api/integrations/google/disconnect:
    delete:
      tags: [Google Integration]
      summary: Disconnect active Google Drive integration
      operationId: deleteGoogleDisconnect
      responses:
        '200':
          description: Integration disconnected or already disconnected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoogleDisconnectResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /api/integrations/google/files:
    get:
      tags: [Google Integration]
      summary: List files from connected Google Drive account
      operationId: getGoogleFiles
      parameters:
        - name: page_size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
        - name: page_token
          in: query
          required: false
          schema:
            type: string
        - name: q
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Google Drive files list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoogleFilesResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          description: Unauthorized or reconnect required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '502':
          $ref: '#/components/responses/GoogleGatewayError'

  /api/folders/tree:
    get:
      tags: [Folders]
      summary: Get folder tree for sidebar
      operationId: getFolderTree
      responses:
        '200':
          description: Folder tree
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FolderTreeResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /api/folders:
    post:
      tags: [Folders]
      summary: Create folder
      operationId: createFolder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFolderRequest'
      responses:
        '201':
          description: Folder created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResource'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  /api/items:
    get:
      tags: [Items]
      summary: List folder content with breadcrumbs
      operationId: listItems
      parameters:
        - name: parent_id
          in: query
          required: false
          schema:
            type: [string, 'null']
        - name: sort_by
          in: query
          required: false
          schema:
            type: string
            enum: [name, type, size, imported_at]
        - name: sort_order
          in: query
          required: false
          schema:
            type: string
            enum: [asc, desc]
      responses:
        '200':
          description: Folder content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListItemsResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  /api/items/{item_id}:
    get:
      tags: [Items]
      summary: Get one Data Room item
      operationId: getItem
      parameters:
        - $ref: '#/components/parameters/ItemId'
      responses:
        '200':
          description: Item metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResource'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    delete:
      tags: [Items]
      summary: Soft-delete item (recursive for folder)
      operationId: deleteItem
      parameters:
        - $ref: '#/components/parameters/ItemId'
      responses:
        '200':
          description: Item deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteItemResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  /api/items/{item_id}/content:
    get:
      tags: [Items]
      summary: Get file content for browser preview/download
      operationId: getItemContent
      parameters:
        - $ref: '#/components/parameters/ItemId'
      responses:
        '200':
          description: File stream
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '410':
          description: File metadata exists but local content is missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /api/items/{item_id}/move:
    patch:
      tags: [Items]
      summary: Move one item
      operationId: moveItem
      parameters:
        - $ref: '#/components/parameters/ItemId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveItemRequest'
      responses:
        '200':
          description: Updated item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResource'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'

  /api/items/{item_id}/copy:
    post:
      tags: [Items]
      summary: Copy one item (recursive for folder)
      operationId: copyItem
      parameters:
        - $ref: '#/components/parameters/ItemId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CopyItemRequest'
      responses:
        '201':
          description: Copied item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResource'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  /api/items/bulk-move:
    post:
      tags: [Items]
      summary: Move multiple items atomically
      operationId: bulkMoveItems
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkMoveRequest'
      responses:
        '200':
          description: Updated items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkItemsResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'

  /api/items/bulk-copy:
    post:
      tags: [Items]
      summary: Copy multiple items atomically
      operationId: bulkCopyItems
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkCopyRequest'
      responses:
        '201':
          description: Copied items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkItemsResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  /api/items/bulk-delete:
    post:
      tags: [Items]
      summary: Delete multiple items atomically
      operationId: bulkDeleteItems
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkDeleteRequest'
      responses:
        '200':
          description: Deleted items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkDeleteResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  /api/files/import-from-google:
    post:
      tags: [Files]
      summary: Import one Google Drive file into Data Room storage
      operationId: postImportFileFromGoogle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportFromGoogleRequest'
      responses:
        '201':
          description: File imported
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResource'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '413':
          description: File too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '502':
          $ref: '#/components/responses/GoogleGatewayError'

  /api/files/upload:
    post:
      tags: [Files]
      summary: Upload one local file into Data Room
      operationId: uploadLocalFile
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                target_folder_id:
                  type: [string, 'null']
              required: [file]
      responses:
        '201':
          description: File uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResource'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '413':
          description: File too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    ItemId:
      name: item_id
      in: path
      required: true
      schema:
        type: string

  responses:
    BadRequestError:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    UnauthorizedError:
      description: Missing/invalid Firebase token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            unauthorized:
              value:
                error:
                  code: unauthorized
                  message: Missing Bearer token.
    ForbiddenError:
      description: Request is authenticated but forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ConflictError:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    GoogleGatewayError:
      description: Upstream Google API request failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'

  schemas:
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          const: ok
      required: [status]

    MeResponse:
      type: object
      properties:
        id:
          type: string
        firebase_uid:
          type: string
        email:
          type: [string, 'null']
          format: email
        display_name:
          type: [string, 'null']
        photo_url:
          type: [string, 'null']
          format: uri
      required: [id, firebase_uid, email, display_name, photo_url]

    GoogleStatusResponse:
      oneOf:
        - type: object
          properties:
            connected:
              type: boolean
              const: false
          required: [connected]
        - type: object
          properties:
            connected:
              type: boolean
              const: true
            google_email:
              type: [string, 'null']
              format: email
            token_expired:
              type: boolean
          required: [connected, google_email, token_expired]

    GoogleConnectResponse:
      type: object
      properties:
        auth_url:
          type: string
          format: uri
      required: [auth_url]

    GoogleDisconnectResponse:
      type: object
      properties:
        connected:
          type: boolean
          const: false
      required: [connected]

    GoogleDriveFile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        mime_type:
          type: [string, 'null']
        size_bytes:
          type: [integer, 'null']
          minimum: 0
        modified_at:
          type: [string, 'null']
          format: date-time
        web_view_link:
          type: [string, 'null']
          format: uri
      required: [id, name, mime_type, size_bytes, modified_at, web_view_link]

    GoogleFilesResponse:
      type: object
      properties:
        files:
          type: array
          items:
            $ref: '#/components/schemas/GoogleDriveFile'
        next_page_token:
          type: [string, 'null']
      required: [files, next_page_token]

    FolderTreeNode:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        children:
          type: array
          items:
            $ref: '#/components/schemas/FolderTreeNode'
      required: [id, name, children]

    FolderTreeResponse:
      type: object
      properties:
        root:
          $ref: '#/components/schemas/FolderTreeNode'
      required: [root]

    ItemResource:
      type: object
      properties:
        id:
          type: string
        kind:
          type: string
          enum: [folder, file]
        name:
          type: string
        parent_id:
          type: [string, 'null']
        status:
          type: string
          enum: [active, failed, deleted]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        mime_type:
          type: [string, 'null']
        size_bytes:
          type: [integer, 'null']
          minimum: 0
        imported_at:
          type: [string, 'null']
          format: date-time
        origin:
          type: [string, 'null']
          enum: [google_drive, local_upload, copied, null]
        google_file_id:
          type: [string, 'null']
      required: [id, kind, name, parent_id, status, created_at, updated_at]

    FolderRef:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        parent_id:
          type: [string, 'null']
      required: [id, name, parent_id]

    Breadcrumb:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
      required: [id, name]

    ListItemsResponse:
      type: object
      properties:
        folder:
          $ref: '#/components/schemas/FolderRef'
        breadcrumbs:
          type: array
          items:
            $ref: '#/components/schemas/Breadcrumb'
        items:
          type: array
          items:
            $ref: '#/components/schemas/ItemResource'
      required: [folder, breadcrumbs, items]

    CreateFolderRequest:
      type: object
      properties:
        parent_id:
          type: [string, 'null']
        name:
          type: string
          minLength: 1
      required: [name]

    ImportFromGoogleRequest:
      type: object
      properties:
        google_file_id:
          type: string
          minLength: 1
        target_folder_id:
          type: [string, 'null']
      required: [google_file_id]

    MoveItemRequest:
      type: object
      properties:
        target_folder_id:
          type: [string, 'null']
      required: [target_folder_id]

    CopyItemRequest:
      type: object
      properties:
        target_folder_id:
          type: [string, 'null']
      required: [target_folder_id]

    BulkMoveRequest:
      type: object
      properties:
        item_ids:
          type: array
          items:
            type: string
          minItems: 1
        target_folder_id:
          type: [string, 'null']
      required: [item_ids, target_folder_id]

    BulkCopyRequest:
      type: object
      properties:
        item_ids:
          type: array
          items:
            type: string
          minItems: 1
        target_folder_id:
          type: [string, 'null']
      required: [item_ids, target_folder_id]

    BulkDeleteRequest:
      type: object
      properties:
        item_ids:
          type: array
          items:
            type: string
          minItems: 1
      required: [item_ids]

    DeleteItemResponse:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
          const: deleted
      required: [id, status]

    BulkItemsResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ItemResource'
      required: [items]

    BulkDeleteResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DeleteItemResponse'
      required: [items]

    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required: [code, message]
      required: [error]